Move & update resumable functionality
